requirements.txt
accelerate==0.27.0
annotated-types==0.6.0
anyio==4.3.0
aspose-cells==24.2.0
bidict==0.22.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip==1.0
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
Flask-SocketIO==5.3.6
fonttools==4.48.1
fsspec==2024.2.0
ftfy==6.1.3
gensim==4.3.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
JPype1==1.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
openai==1.13.3
packaging==23.2
pandas==2.2.0
pillow==10.2.0
psutil==5.9.8
pydantic==2.6.3
pydantic_core==2.16.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sympy==1.12
threaded==4.2.0
threadpoolctl==3.2.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
validators==0.22.0
wcwidth==0.2.13
Werkzeug==3.0.1
wsproto==1.2.0
